requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
Authlib==1.3.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gritql==0.1.5
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-cli==0.0.35
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langserve==0.3.1
langsmith==0.2.10
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.1
openai==1.59.4
orjson==3.10.13
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==1.8.2
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
weaviate-client==4.10.2
yarl==1.18.3
